refactor(cmd): move expand helpers from create.go to expand.go
Five helpers were annotated "// TODO: move to expand.go". They support
blueprint expansion (validation level, validator skipping, CLI variable
and backend overrides, and post-expand validation) and are used by
expandOrDie, which both the create and expand commands invoke. Move them
to expand.go and drop the now-unused imports (errors, validators, cty,
yaml) from create.go.

Pure relocation, no behavior change. go build, go vet, gofmt and
go test ./cmd/... all pass.
